Provide a helpful error on a NULL PK value.
This will typically mean that the end-user specified the wrong PK COLUMN
so we want to be helpful about it.  We have to retrieve the "IS NULL"
status for the PK attribute so this check does not add any overhead over
a little bit check. (-: We only retrieve the PK, TABLE NAME and other
information iff the error is encountered and in that case we want to be
as helpful as possible.