Add DELETE support while factoring out common code.
After the planning and initial setup in the earlier multiple callback
functions, the final stage of UPDATE and DELETE is nearly identical so
we factor out the common code.

While we are at it, also expand the conditional compilation scope for
the Write API to the utility functions as well as adding basic function
header comments.